sucessfully converted level11.json and level12.json to updated level struct
created Enemy1, 2, Level1, 2 for versioning structs. the plan is to do this whenever a struct is updated. ex: Item 1, Item 2.
version.odin handles the conversion from different versions of a struct as well as converting existing json files from one version of a struct to another. at the moment the process for converting a save file is 1) fix any issues with utility.py, then 2) run the test function inside version.odin with the proper file names and converter procedure. In the previous two commits I also made various fixes that came to mind: resetting idle_look_timer at start of idle state, moving the enemy.post_pos = enemy.pos into the setup_enemy function as opposed to load_level.

further improvements: create a better system for converting level saves and other saved data. Perhaps store the version in the file at the top and then automatically convert them in their corresponding load function. convert functions will not allocate new data, if it is not necessary this could be bad, but in our case in the way these functions are being used I don't think it will be