editor improvments and a spawn state for enemies
next steps are just to make some levels with this. add a itty bit of polish and then add some upgrades in. combat is feeling okay right now. sort of feels all or nothing. either your bashing them or there's too much distance and they kill you. arrows are to avoid and melee enemy is hard to avoid too